CUDA support started

+++ ggml/src/ggml-cpu/quants.c
@@ -614,7 +614,8 @@ void ggml_vec_dot_q3_k_hifi_q8_K_generic(int n, float * GGML_RESTRICT s, size_t 
             // We need to subtract the ~0 Q3_K contribution and add the original outlier value
             for (int k = 0; k < n_out; ++k) {
                 uint8_t idx = xb->outlier_idx[k];
-                if (idx < Q3_K_HIFI_BLOCK_SIZE) {
+                // Cast to int to avoid tautological comparison warning (uint8_t is always < 256)
+                if ((int)idx < Q3_K_HIFI_BLOCK_SIZE) {
                     float outlier_val = GGML_FP16_TO_FP32(xb->outliers[k]);
                     float q3k_val = q3k_weights[idx];  // Should be ~0 since we zeroed it
                     q3k_sum += (outlier_val - q3k_val) * (float)q8[idx] * d_y;
